requirements.txt
Flask==1.1.1
line-bot-sdk==1.13.0
urllib3==1.13.0
beautifulsoup==4.7.
soupsieve==1.6.1